requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
greenlet==3.2.3
idna==3.10
numpy==2.3.2
pandas==2.3.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
